integration: move *test to internal/
...primarily to make it easier to identify which subpackages are actual
test suites.
