+++ integration/internal/vaulttest/cmd.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/concourse/concourse/integration/cmdtest"
-	"github.com/concourse/concourse/integration/dctest"
+	"github.com/concourse/concourse/integration/internal/cmdtest"
+	"github.com/concourse/concourse/integration/internal/dctest"
 	"github.com/stretchr/testify/require"
 )
