+++ integration/ops/upgrade_test.go
@@ -3,8 +3,8 @@ package ops_test
 import (
 	"testing"
 
-	"github.com/concourse/concourse/integration/dctest"
-	"github.com/concourse/concourse/integration/flytest"
+	"github.com/concourse/concourse/integration/internal/dctest"
+	"github.com/concourse/concourse/integration/internal/flytest"
 	"github.com/stretchr/testify/require"
 )
